+++ packages/core/src/index.ts
@@ -1,4 +1,5 @@
 export type * from './components.d.ts';
+import { ExcelDefinition } from './declarations/ExcelDefinition.js';
 import { importExcel, exportExcel, downloadExcelTemplate, fromExcel, toExcel, generateExcelTemplate, initializeWasm } from './utils.js';
 
 function getUtils() {
@@ -13,4 +14,24 @@ function getUtils() {
     }
 }
 
-export { getUtils };
+function _importExcel<T>(defintion: ExcelDefinition): Promise<T[]> {
+    initializeWasm();
+    return importExcel(defintion);
+}
+
+function _exportExcel<T>(defintion: ExcelDefinition, data: T[]): Promise<void> {
+    initializeWasm();
+    return exportExcel(defintion, data);
+}
+
+function _downloadExcelTemplate(defintion: ExcelDefinition): Promise<void> {
+    initializeWasm();
+    return downloadExcelTemplate(defintion);
+}
+
+export {
+    getUtils,
+    _importExcel as importExcel,
+    _exportExcel as exportExcel,
+    _downloadExcelTemplate as downloadExcelTemplate
+};