+++ packages/wasm/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @senlinz/import-export-wasm
 
+## 0.1.0-beta.8
+
+### Patch Changes
+
+- add wrapper function in index.js, fix definition datetime undefined
+
 ## 0.1.0-beta.7
 
 ### Minor Changes